+++ SilKit/source/services/orchestration/LifecycleManagement.cpp
@@ -251,6 +251,14 @@ void LifecycleManagement::StartTime()
     (dynamic_cast<TimeSyncService*>(_lifecycleService->GetTimeSyncService()))->StartTime();
 }
 
+
+void LifecycleManagement::StopTime()
+{
+    (dynamic_cast<TimeSyncService*>(_lifecycleService->GetTimeSyncService()))->StopTime();
+}
+
+
+
 void LifecycleManagement::AddAsyncSubscriptionsCompletionHandler(std::function<void()> handler)
 {
     _lifecycleService->AddAsyncSubscriptionsCompletionHandler(std::move(handler));