+++ SilKit/source/services/orchestration/LifecycleManagement.hpp
@@ -68,6 +68,7 @@ public: //CTors
 
     // Send NextSimTask
     void StartTime();
+    void StopTime();
 
     // Callback handling
     CallbackResult HandleCommunicationReady();