+++ SilKit/source/services/orchestration/TimeSyncService.hpp
@@ -52,7 +52,6 @@ class TimeSyncService
     , public IMsgForTimeSyncService
     , public Core::IServiceEndpoint
 {
-    friend struct DistributedTimeQuantumPolicy;
 
 public:
     // ----------------------------------------
@@ -61,7 +60,9 @@ public:
     // ----------------------------------------
     // Constructors, Destructor, and Assignment
     TimeSyncService(Core::IParticipantInternal* participant, ITimeProvider* timeProvider,
-                    const Config::HealthCheck& healthCheckConfig, LifecycleService* lifecycleService);
+                    const Config::HealthCheck& healthCheckConfig, LifecycleService* lifecycleService, double animationFactor = 0);
+
+    ~TimeSyncService();
 
 public:
     // ----------------------------------------
@@ -85,7 +86,8 @@ public:
     void ResetTime();
     void ConfigureTimeProvider(Orchestration::TimeProviderKind timeProviderKind);
     void StartTime();
-
+    void StopTime();
+    
     // IServiceEndpoint
     inline void SetServiceDescriptor(const Core::ServiceDescriptor& serviceDescriptor) override;
     inline auto GetServiceDescriptor() const -> const Core::ServiceDescriptor& override;
@@ -103,6 +105,9 @@ public:
 
     void RequestNextStep();
 
+    auto IsCoupledToWallClock() const -> bool;
+    auto GetCurrentWallClockSyncPoint() const -> std::chrono::nanoseconds;
+
 private:
     // ----------------------------------------
     // private methods
@@ -112,6 +117,9 @@ private:
     bool SetupTimeSyncPolicy(bool isSynchronizingVirtualTime);
 
     inline auto GetTimeSyncPolicy() const -> ITimeSyncPolicy*;
+    
+    void StopWallClockCouplingThread();
+    void StartWallClockCouplingThread();
 
 private:
     // ----------------------------------------
@@ -138,6 +146,17 @@ private:
     Util::PerformanceMonitor _execTimeMonitor;
     Util::PerformanceMonitor _waitTimeMonitor;
     WatchDog _watchDog;
+    bool _isCoupledToWallClock{false}; 
+    std::thread _wallClockCouplingThread;
+    std::promise<void> _wallClockCouplingThreadStopPromise;
+    std::future<void> _stopFuture;
+    mutable std::mutex _mx;
+    using Lock = std::unique_lock<decltype(_mx)>;
+    std::chrono::nanoseconds _currentWallClockSyncPoint{0ns};
+    double _animationFactor{0};
+    std::atomic<bool> _wallClockCouplingThreadRunning{false};
+    std::atomic<bool> _wallClockReachedBeforeCompletion{false};
+    
 };
 
 // ================================================================================